refactor(interface): split soroban.ts, consolidate wallet store
Splits the 616-line lib/soroban.ts (chain reads already moved out in
#1053) into lib/soroban/{client,tx-builders,submit}.ts, with a memoized
Horizon/RPC client shared across calls instead of a fresh instance per
call. Collapses store/useWalletStore.ts + store/walletSlice.ts +
hooks/useWalletSlice.ts into a single canonical store module, keeping
useWalletSlice as the one thin hook layer. Public import paths and the
useWalletSlice/useWalletStatus/useWalletSigning hook API are unchanged.

Also fixes several pre-existing issues found while getting CI green:
duplicate validateCampaignInput/validateDonationAmount declarations in
packages/types (broke the @fund-my-cause/sdk build for every PR),
missing typecheck scripts in apps/interface and apps/components-lib,
jest.config.ts requiring an unavailable ts-node (converted to .mjs),
and dead/incomplete component extractions in ReferralProgram.tsx and
TransactionHistory.tsx that failed typecheck.
