+++ apps/components-lib/package.json
@@ -62,6 +62,7 @@
   "scripts": {
     "build": "tsc",
     "dev": "tsc --watch",
+    "typecheck": "tsc --noEmit",
     "lint": "eslint src",
     "test": "vitest"
   },
