+++ apps/interface/src/components/ui/TransactionHistory.tsx
@@ -11,6 +11,8 @@ import {
   NoTransactionsIllustration,
 } from "@/components/ui/EmptyState";
 import { TransactionExportModal } from "@/components/ui/TransactionExportModal";
+import { TransactionTable } from "@/components/ui/transaction/TransactionTable";
+import { TransactionFilterPanel } from "@/components/ui/transaction/TransactionFilterPanel";
 import {
   applyFilters,
   type ExportRecord,
@@ -184,92 +186,21 @@ export function TransactionHistory({
         </div>
 
         {showFilters && (
-          <div className="flex flex-wrap gap-3 p-3 bg-gray-50 dark:bg-gray-800/60 rounded-xl border border-gray-200 dark:border-gray-700">
-            {/* Type filter */}
-            <div className="flex flex-col gap-1">
-              <label className="text-xs text-gray-500 dark:text-gray-400">
-                Type
-              </label>
-              <select
-                value={typeFilter}
-                onChange={(e) => setTypeFilter(e.target.value)}
-                className={inputCls}
-                aria-label="Filter by type"
-              >
-                <option value="">All types</option>
-                <option value="Confirmed">Confirmed</option>
-                <option value="Pending">Pending</option>
-                <option value="Failed">Failed</option>
-              </select>
-            </div>
-
-            {/* Date range */}
-            <div className="flex flex-col gap-1">
-              <label className="text-xs text-gray-500 dark:text-gray-400">
-                From
-              </label>
-              <input
-                type="date"
-                value={dateRange.from}
-                onChange={(e) =>
-                  setDateRange((p) => ({ ...p, from: e.target.value }))
-                }
-                className={inputCls}
-                aria-label="Filter from date"
-              />
-            </div>
-            <div className="flex flex-col gap-1">
-              <label className="text-xs text-gray-500 dark:text-gray-400">
-                To
-              </label>
-              <input
-                type="date"
-                value={dateRange.to}
-                min={dateRange.from || undefined}
-                onChange={(e) =>
-                  setDateRange((p) => ({ ...p, to: e.target.value }))
-                }
-                className={inputCls}
-                aria-label="Filter to date"
-              />
-            </div>
-
-            {/* Campaign filter (shown when multiple campaigns could appear) */}
-            {campaignIds.length > 1 && (
-              <div className="flex flex-col gap-1">
-                <label className="text-xs text-gray-500 dark:text-gray-400">
-                  Campaign
-                </label>
-                <select
-                  value={campaignFilter}
-                  onChange={(e) => setCampaignFilter(e.target.value)}
-                  className={inputCls}
-                  aria-label="Filter by campaign"
-                >
-                  <option value="">All campaigns</option>
-                  {campaignIds.map((id) => (
-                    <option key={id} value={id}>
-                      {id.slice(0, 8)}…
-                    </option>
-                  ))}
-                </select>
-              </div>
-            )}
-
-            {/* Clear */}
-            {isFiltered && (
-              <button
-                onClick={() => {
-                  setTypeFilter("");
-                  setDateRange({ from: "", to: "" });
-                  setCampaignFilter("");
-                }}
-                className="self-end text-xs text-indigo-500 hover:underline"
-              >
-                Clear filters
-              </button>
-            )}
-          </div>
+          <TransactionFilterPanel
+            typeFilter={typeFilter}
+            onTypeFilterChange={setTypeFilter}
+            dateRange={dateRange}
+            onDateRangeChange={setDateRange}
+            campaignFilter={campaignFilter}
+            onCampaignFilterChange={setCampaignFilter}
+            campaignIds={campaignIds}
+            isFiltered={isFiltered}
+            onClearFilters={() => {
+              setTypeFilter("");
+              setDateRange({ from: "", to: "" });
+              setCampaignFilter("");
+            }}
+          />
         )}
 
         {isFiltered && (
@@ -282,64 +213,10 @@ export function TransactionHistory({
         )}
 
         {/* Table */}
-        {filteredRecords.length === 0 ? (
-          <p className="text-sm text-gray-500 dark:text-gray-400 text-center py-6">
-            No contributions match your filters.
-          </p>
-        ) : (
-          <div className="rounded-xl border border-gray-200 dark:border-gray-800 overflow-hidden">
-            <table className="w-full text-sm">
-              <thead>
-                <tr className="bg-gray-100 dark:bg-gray-800 text-gray-500 dark:text-gray-400 text-xs uppercase tracking-wide">
-                  <th className="px-4 py-2 text-left font-medium">
-                    Contributor
-                  </th>
-                  <th className="px-4 py-2 text-right font-medium">Amount</th>
-                  <th className="px-4 py-2 text-right font-medium">Date</th>
-                  <th
-                    className="px-4 py-2 text-right font-medium"
-                    aria-label="View transaction link"
-                  >
-                    <span className="sr-only">Link</span>
-                  </th>
-                </tr>
-              </thead>
-              <tbody className="divide-y divide-gray-100 dark:divide-gray-800">
-                {displayRecords.map((r) => (
-                  <tr
-                    key={r.txHash}
-                    className="bg-white dark:bg-gray-900 hover:bg-gray-50 dark:hover:bg-gray-800/60 transition-colors"
-                  >
-                    <td className="px-4 py-3 font-mono text-gray-700 dark:text-gray-300">
-                      <span title={r.contributor}>
-                        {truncate(r.contributor)}
-                      </span>
-                    </td>
-                    <td className="px-4 py-3 text-right text-gray-900 dark:text-white font-medium">
-                      {r.amountXlm > 0
-                        ? `${r.amountXlm.toLocaleString(undefined, { maximumFractionDigits: 7 })} XLM`
-                        : "—"}
-                    </td>
-                    <td className="px-4 py-3 text-right text-gray-500 dark:text-gray-400">
-                      {formatDate(r.timestamp)}
-                    </td>
-                    <td className="px-4 py-3 text-right">
-                      <a
-                        href={`${STELLAR_EXPERT}/tx/${r.txHash}`}
-                        target="_blank"
-                        rel="noopener noreferrer"
-                        aria-label="View transaction on Stellar Expert"
-                        className="inline-flex items-center text-indigo-500 hover:text-indigo-400"
-                      >
-                        <ExternalLink size={14} />
-                      </a>
-                    </td>
-                  </tr>
-                ))}
-              </tbody>
-            </table>
-          </div>
-        )}
+        <TransactionTable
+          records={displayRecords}
+          stellarExpertBaseUrl={STELLAR_EXPERT}
+        />
 
         {filteredRecords.length > 10 && (
           <p className="text-xs text-gray-400 dark:text-gray-500 text-right">
